+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fec-builder",
-  "version": "2.0.0-beta.10",
+  "version": "2.0.0-beta.11",
   "bin": {
     "fec-builder": "./lib/bin.js"
   },
@@ -28,7 +28,7 @@
     "css-loader": "^5.0.1",
     "css-minimizer-webpack-plugin": "^1.3.0",
     "del": "^6.0.0",
-    "html-webpack-plugin": "^5.0.0-alpha.14",
+    "html-webpack-plugin": "5.3.2",
     "immer": "^8.0.3",
     "less": "^3.12.2",
     "less-loader": "^7.1.0",
@@ -40,11 +40,12 @@
     "qiniu": "^7.3.2",
     "react-refresh": "^0.9.0",
     "semver": "^7.3.2",
+    "source-map-loader": "^3.0.0",
     "style-loader": "^2.0.0",
-    "ts-loader": "^8.0.11",
+    "ts-loader": "^9.2.4",
     "typescript": "~4.1.3",
     "walk": "^2.3.14",
-    "webpack": "^5.4.0",
+    "webpack": "^5.13.0",
     "webpack-bundle-analyzer": "^4.4.1",
     "webpack-dev-server": "^3.11.0",
     "webpackbar": "^5.0.0-3",
