+++ src/schema2md.ts
@@ -163,7 +163,7 @@ export default function main() {
   const schema = require('../preset-configs/config.schema.json')
   const markdown = schemaToMarkdown(schema, { level: 0, keyPath: [], title: 'Build Config' })
 
-  const targetPath = path.join(__dirname, './build-config.md')
+  const targetPath = path.resolve(__dirname, '../build-config.md')
   console.log('Markdown Output Path: ' + targetPath)
 
   try {
